Add files via upload
Establish a class hierarchy, using the base class called "Beverage" and the derived classes called "Coffee," "Tea," "Soda," and "EnergyDrink." Together with particular properties for the derived classes, each class should include the following attributes: Name, Description, Serving Size, Calories, and Price. To describe each beverage, provide setter/getter methods, parameterized constructors, destructors, and a toString() function. Create a software that makes advantage of the parameterized constructors to let users enter data for different types of drinks. Make a list or vector of these drinks using a loop, asking the user for pertinent details. Lastly, present the assembled drinks. Certain properties, including the hot/cold nature of coffee, the kind of tea, the brand of soda, and the stimulant in energy drinks, should be included in the derived classes.